tests 1-5 work

+++ src/entities/audio/Song.java
@@ -14,7 +14,7 @@ import java.util.Iterator;
 public final class Song implements Audio {
     private final SongInput songInput;
     @Getter
-    private float revenue = 0;
+    private double revenue = 0;
     private final ArrayList<NormalUser> likedBy = new ArrayList<>();
     /**
      * -- GETTER --
@@ -313,7 +313,7 @@ public final class Song implements Audio {
      *
      * @param amount The amount of money to addd
      */
-    public void addRevenue(float amount) {
+    public void addRevenue(double amount) {
         revenue += amount;
     }
 }
